pkg/kfuzztest: supported nested structures in DWARF parsing logic
The DWARF parsing logic was failing to handle cases where a structure
contained a pointer to another structure or a qualified pointer (e.g.,
const struct *mystruct, which contains two levels of indirection from a
DWARF perspective). The update allows us to extract these also, which is
required for generating valid syzlang descriptions for KFuzzTest
targets.