syz-kfuzztest: add standalone tool for in-vm KFuzzTest fuzzing
We add a rudimentary tool for developers to use to fuzz KFuzzTest
targets inside a VM. We make use of most of the other infrastructure built
around KFuzzTest, including but not limited to the specialized encoding
encoding and automatic target extraction from a vmlinux binary. This
tool is written in pure Go with no depdenency on the C++ executor codebase,
using only Go's filesystem API for sending KFuzzTest inputs into the
kernel.

This tool is still limited by the fact that it doesn't have support for
KCOV, as this will require writing a new library for doing this.

kfuzztest standalone: