pkg/prog: update marshalling for kfuzztest input arguments
The current relocation table format would pack everything into a single
kernel heap allocation during deserialization. The problem with this is
that if a some data doesn't sit at the boundary of this region, we
cannot detect buffer overflows on it.

Every logical node in the argument graph now gets its own heap-allocated
memory region, that is exactly the right size. Even if the kernel
allocated more memory for it during kmalloc, the boundaries will still
be poisoned and as a result we should be able to catch KASAN buffer
overflows on all data.

This update involves adding new metadata to the KFuzzTest input, namely
a region array which represents the size and alignment characteristics
of every piece of data. Other updates are made to the relocation table
serialization, where now instead of storing relative offsets in order to
compute the pointer's address, we store a region identifier to express
that the pointer points to the beginning of a region (some other
argument).