prog/target: add fetching function for syz_kfuzztest_run ID
All non-base variants of syz_kfuzztest_run (i.e., those that are
discovered dynamically) are encoded so that they map onto the base
variant which is defined in kfuzztest.txt, and known by the executor.

We add a function for fetching this, that is wrapped in a sync.once
block to avoid repeated iteration over the target's array of syscalls.

Signed-off-by: Ethan Graham <ethangraham@google.com>